Fix detection of system `node` command in binary
Because of `pkg`, we have to specify where we're running from and break
out of the `pkg` snapshot file system. Using `/usr/bin/env` like you
would in a script works perfectly. Only downside is that this will fail
on non-Unix-like systems, but with the current limitation to Bash that
isn't really an issue.